agents/requirements.txt
fastapi==0.115.6
fastapi-cli==0.0.7
llama-cloud==0.1.6
llama-index==0.12.8
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.8
llama-index-embeddings-azure-openai==0.3.0
llama-index-embeddings-gemini==0.3.0
llama-index-embeddings-openai==0.3.1
llama-index-embeddings-ollama==0.6.0
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-azure-openai==0.3.0
llama-index-llms-gemini==0.4.2
llama-index-llms-openai==0.3.12
llama-index-llms-ollama==0.5.3
llama-index-multi-modal-llms-openai==0.4.1
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-database==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-lancedb==0.3.0
llama-parse==0.5.18
lancedb==0.17.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic[email]
python-multipart==0.0.20
SQLAlchemy==2.0.36
uvicorn==0.34.0
aio-pika==9.4.0
